re-arranged methods
