+++ Gemfile
@@ -1,3 +1,4 @@
 gem 'rspec'
 gem 'rubocop'
+gem 'colorize'
 source 'https://rubygems.org'
