--- gem
@@ -1,1 +1,0 @@
-Gem 'Rubocop'
