+++ lib/help.rb
@@ -1,0 +1,7 @@
+def __help
+    helping = "This program is a linter for Gemfiles in ruby.
+  You might say well Ruby doesn't need Gemfile linter,
+  buy Hey! nothing is an overkill if it can do the work."
+  
+    puts helping
+end
